requirements.txt

streamlit
google-generativeai
json